refactor(extensions): remove coding-agent extension/skills shim modules
