--- crates/tau-coding-agent/src/skills.rs
@@ -1,1 +1,0 @@
-pub use tau_skills::*;
