+++ crates/tau-extensions/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 anyhow.workspace = true
 serde.workspace = true
 serde_json.workspace = true
+tau-cli = { path = "../tau-cli" }
 wait-timeout.workspace = true
 
 [dev-dependencies]
